Add test case for express app
