+++ package.json
@@ -10,11 +10,19 @@
   },
   "author": "Sravya Padala",
   "license": "ISC",
+  "jest": {
+    "verbose": true,
+    "globals": {
+      "__DEV__": true,
+      "__RCTProfileIsProfiling": false
+    }
+  },
   "devDependencies": {
     "@babel/core": "^7.23.9",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/preset-env": "^7.23.9",
-    "@types/jest": "^29.5.12",
     "babel-jest": "^29.7.0",
+    "babel-plugin-styled-components": "^2.1.4",
     "eslint": "^8.56.0",
     "eslint-config-standard": "^17.1.0",
     "eslint-plugin-import": "^2.29.1",
@@ -25,7 +33,6 @@
     "supertest": "^6.3.4"
   },
   "dependencies": {
-    "express": "^4.18.2",
-    "ts-jest": "^29.1.2"
+    "express": "^4.18.2"
   }
 }
