+++ server.js
@@ -1,16 +1,3 @@
-import express from 'express'
-import userRouter from './routes/users.js'
-import adminRouter from './routes/admin.js'
+import app from './app.js'
 
-const app = express()
 app.listen(3000)
-app.get('/', (req, res) => {
-  res.send('Running Okay')
-})
-
-// const userRouter = require('./routes/users')
-
-app.use('/users', userRouter)
-app.use('/admin', adminRouter)
-
-export default app
